+++ mysql_db_dump.py
@@ -8,7 +8,8 @@
 
     Usage:
         mysql_db_dump.py -c file -d path {-B db_name [db_name ...] |
-            -A | -D} [-o name | -p path | -s | -z | -r] [-v | -h]
+            -A | -D} [-o name | -p path | -s | -z | -r] [-y flavor_id]
+            [-v | -h]
 
     Arguments:
         -B databases => Database names, space delimited.
@@ -22,7 +23,8 @@
             in the $PATH variable.)
         -s => Run dump as a single transaction.
         -r => Remove GTID entries from dump file.
-        -z => Compress database dump files
+        -z => Compress database dump files.
+        -y value => A flavor id for the program lock.  To create unique lock.
         -v => Display version of this program.
         -h => Help and usage message.
 
@@ -30,8 +32,8 @@
         NOTE 2:  -A, -B, and -D are XOR arguments.
 
     Notes:
-        Database configuration file format (mysql_{host}.py):
-            # Configuration file for {Database Name/Server}
+        Database configuration file format (mysql_cfg.py.TEMPLATE):
+            # Configuration file for Database
             user = "root"
             passwd = "ROOT_PASSWORD"
             host = "IP_ADDRESS"
@@ -42,13 +44,19 @@
             sid = "SERVER_ID"
             extra_def_file = "DIRECTORY_PATH/myextra.cfg"
 
-        NOTE:  Include the cfg_file even if running remotely as the file will
+        NOTE 1:  Include the cfg_file even if running remotely as the file will
             be used in future releases.
+
+        NOTE 2:  In MySQL 5.6 - it now gives warning if password is passed on
+            the command line.  To suppress this warning, will require the use
+            of the --defaults-extra-file option (i.e. extra_def_file) in the
+            database configuration file.  See below for the defaults-extra-file
+            format.
 
         configuration modules -> name is runtime dependent as it can be
             used to connect to different databases with different names.
 
-        Defaults Extra File format (filename.cfg):
+        Defaults Extra File format (mysql.cfg.TEMPLATE):
             [client]
             password="ROOT_PASSWORD"
             socket="DIRECTORY_PATH/mysql.sock"
@@ -69,12 +77,12 @@ import sys
 # Local
 import lib.arg_parser as arg_parser
 import lib.gen_libs as gen_libs
+import lib.gen_class as gen_class
 import lib.cmds_gen as cmds_gen
 import mysql_lib.mysql_class as mysql_class
 import mysql_lib.mysql_libs as mysql_libs
 import version
 
-# Version
 __version__ = version.__version__
 
 
@@ -92,14 +100,14 @@ def help_message():
     print(__doc__)
 
 
-def crt_dump_cmd(SERVER, args_array, opt_arg_list, opt_dump_list):
+def crt_dump_cmd(server, args_array, opt_arg_list, opt_dump_list, **kwargs):
 
     """Function:  crt_dump_cmd
 
     Description:  Create the database dump command line.
 
     Arguments:
-        (input) SERVER -> Database server instance.
+        (input) server -> Database server instance.
         (input) args_array -> Array of command line options and values.
         (input) opt_arg_list -> List of commands to add to cmd line.
         (input) opt_dump_list -> Dictionary of additional options.
@@ -107,9 +115,11 @@ def crt_dump_cmd(SERVER, args_array, opt_arg_list, opt_dump_list):
 
     """
 
-    dump_args = mysql_libs.crt_cmd(SERVER, arg_parser.arg_set_path(args_array,
-                                                                   "-p") +
-                                   "mysqldump")
+    args_array = dict(args_array)
+    opt_dump_list = dict(opt_dump_list)
+    opt_arg_list = list(opt_arg_list)
+    dump_args = mysql_libs.crt_cmd(
+        server, arg_parser.arg_set_path(args_array, "-p") + "mysqldump")
 
     # Add arguments to dump command.
     for arg in opt_arg_list:
@@ -129,18 +139,17 @@ def dump_run(dump_cmd, dmp_file, compress, **kwargs):
         (input) dump_cmd -> Database dump command line.
         (input) compress -> Compression flag.
         (input) dmp_file -> Dump file and path name.
-        (input) **kwargs:
-            None
 
     """
 
+    dump_cmd = list(dump_cmd)
     cmds_gen.run_prog(dump_cmd, ofile=dmp_file)
 
     if compress:
         gen_libs.compress(dmp_file)
 
 
-def dump_db(dump_cmd, db_list, compress, dmp_path):
+def dump_db(dump_cmd, db_list, compress, dmp_path, **kwargs):
 
     """Function:  dump_db
 
@@ -155,12 +164,13 @@ def dump_db(dump_cmd, db_list, compress, dmp_path):
 
     """
 
-    if db_list:
+    dump_cmd = list(dump_cmd)
+    db_list = list(db_list)
 
+    if db_list:
         for db in db_list:
             dump_cmd = cmds_gen.add_cmd(dump_cmd, arg=db)
             dmp_file = gen_libs.crt_file_time(db, dmp_path, ".sql")
-
             dump_run(dump_cmd, dmp_file, compress)
 
             # Remove database name from command.
@@ -168,14 +178,13 @@ def dump_db(dump_cmd, db_list, compress, dmp_path):
 
     elif "--all-databases" in dump_cmd:
         dmp_file = gen_libs.crt_file_time("All_Databases", dmp_path, ".sql")
-
         dump_run(dump_cmd, dmp_file, compress)
 
     else:
         print("WARNING:  No databases to dump or missing -D option.")
 
 
-def set_db_list(SERVER, args_array, **kwargs):
+def set_db_list(server, args_array, **kwargs):
 
     """Function:  set_db_list
 
@@ -183,15 +192,14 @@ def set_db_list(SERVER, args_array, **kwargs):
         database is being selected.
 
     Arguments:
-        (input) SERVER -> Database server instance.
+        (input) server -> Database server instance.
         (input) args_array -> Array of command line options and values.
         (output) -> Database list.
-        (input) **kwargs:
-            multi_val -> List of options that may have multiple values.
 
     """
 
-    db_list = gen_libs.dict_2_list(mysql_libs.fetch_db_dict(SERVER),
+    args_array = dict(args_array)
+    db_list = gen_libs.dict_2_list(mysql_libs.fetch_db_dict(server),
                                    "Database")
 
     # Specified databases.
@@ -223,23 +231,21 @@ def run_program(args_array, opt_arg_list, opt_dump_list, **kwargs):
         (input) args_array -> Array of command line options and values.
         (input) opt_arg_list -> List of commands to add to cmd line.
         (input) opt_dump_list -> Dictionary of additional options.
-        (input) **kwargs:
-            multi_val -> List of options that may have multiple values.
 
     """
 
-    SERVER = mysql_libs.create_instance(args_array["-c"], args_array["-d"],
+    args_array = dict(args_array)
+    opt_dump_list = dict(opt_dump_list)
+    opt_arg_list = list(opt_arg_list)
+    server = mysql_libs.create_instance(args_array["-c"], args_array["-d"],
                                         mysql_class.Server)
-
-    SERVER.connect()
-    SERVER.set_srv_gtid()
-
-    dump_cmd = crt_dump_cmd(SERVER, args_array, opt_arg_list, opt_dump_list)
-
-    db_list = set_db_list(SERVER, args_array, **kwargs)
+    server.connect()
+    server.set_srv_gtid()
+    dump_cmd = crt_dump_cmd(server, args_array, opt_arg_list, opt_dump_list)
+    db_list = set_db_list(server, args_array, **kwargs)
 
     # Remove the -r option if database is not GTID enabled.
-    if "-r" in args_array and not SERVER.gtid_mode:
+    if "-r" in args_array and not server.gtid_mode:
         dump_cmd.remove(opt_dump_list["-r"])
 
     compress = False
@@ -253,8 +259,7 @@ def run_program(args_array, opt_arg_list, opt_dump_list, **kwargs):
         dmp_path = args_array["-o"] + "/"
 
     dump_db(dump_cmd, db_list, compress, dmp_path)
-
-    cmds_gen.disconnect([SERVER])
+    cmds_gen.disconnect([server])
 
 
 def main():
@@ -281,6 +286,7 @@ def main():
 
     dir_chk_list = ["-o", "-d", "-p"]
     dir_crt_list = ["-o"]
+
     # --ignore-table=mysql.event -> Skips dumping the event table.
     opt_arg_list = ["--ignore-table=mysql.event"]
     opt_dump_list = {"-s": "--single-transaction",
@@ -289,20 +295,29 @@ def main():
                      "-r": "--set-gtid-purged=OFF"}
     opt_multi_list = ["-B"]
     opt_req_list = ["-c", "-d"]
-    opt_val_list = ["-B", "-c", "-d", "-o", "-p"]
+    opt_val_list = ["-B", "-c", "-d", "-o", "-p", "-y"]
     opt_xor_dict = {"-A": ["-B", "-D"], "-B": ["-A", "-D"], "-D": ["-A", "-B"]}
 
     # Process argument list from command line.
     args_array = arg_parser.arg_parse2(sys.argv, opt_val_list,
                                        multi_val=opt_multi_list)
 
-    if not gen_libs.help_func(args_array, __version__, help_message):
-        if arg_parser.arg_xor_dict(args_array, opt_xor_dict) \
-           and not arg_parser.arg_require(args_array, opt_req_list) \
-           and not arg_parser.arg_dir_chk_crt(args_array, dir_chk_list,
-                                              dir_crt_list):
+    if not gen_libs.help_func(args_array, __version__, help_message) \
+       and not arg_parser.arg_require(args_array, opt_req_list) \
+       and arg_parser.arg_xor_dict(args_array, opt_xor_dict) \
+       and not arg_parser.arg_dir_chk_crt(args_array, dir_chk_list,
+                                          dir_crt_list):
+
+        try:
+            prog_lock = gen_class.ProgramLock(sys.argv,
+                                              args_array.get("-y", ""))
             run_program(args_array, opt_arg_list, opt_dump_list,
                         multi_val=opt_multi_list)
+            del prog_lock
+
+        except gen_class.SingleInstanceException:
+            print("WARNING:  Lock in place for mysql_db_dump with id: %s"
+                  % (args_array.get("-y", "")))
 
 
 if __name__ == "__main__":